Limit support to Puppet 5
metadata.json actually said this only supported pre-puppet 5... that was
flat wrong and on oversight. This module has been in use on Puppet 5
since Puppet 5 came out. The file has been corrected and not confines
the supported version to just Puppet 5 since everything before that is
EoL now. I am not doing a major version bump as no code that would
impact support of older versions was touched here.